Delegated task use to TaskManager and added progress percentage
